services/kafka_to_feature_store/pyproject.toml
[tool.poetry]
name = "kafka-to-feature-store"
version = "0.1.0"
description = ""
authors = ["Fabio Scramim Rigo <fabio.s.rigo@protonmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
quixstreams = "^2.8.1"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
pydantic-settings = "^2.3"
hopsworks = "^3.7.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
extend-select = ["I"]